requirements.txt
streamlit
pandas
qrcode
python-barcode
Pillow